refactor: Separates navbar learner controls into a standalone component.
