+++ src/app/modules/layout/navbar/learner-controls/learner-controls.component.html
@@ -1,0 +1,15 @@
+<button color="primary" id="unit" mat-flat-button [disabled]="!learner" [matMenuTriggerFor]="unit">
+    {{selectedUnit ? selectedUnit.name : 'Izaberi lekciju'}}
+</button>
+<mat-menu #unit="matMenu">
+<button [routerLink]="['unit', unit.id]" *ngFor="let unit of units" mat-menu-item>
+    {{unit.name}}
+</button>
+</mat-menu>
+<button color="primary" id="kc" mat-flat-button [matMenuTriggerFor]="kc" [disabled]="!selectedUnit">
+    {{selectedKC ? selectedKC.name : 'Izaberi veštinu'}}
+</button>
+<mat-menu #kc="matMenu">
+<cc-knowledge-map *ngIf="selectedUnit" [unitId]="selectedUnit.id" [knowledgeComponents]="selectedUnit.knowledgeComponents">
+</cc-knowledge-map>
+</mat-menu>
