+++ src/app/modules/layout/navbar/learner-controls/learner-controls.component.scss
@@ -1,0 +1,25 @@
+:host {
+    display: block;
+  }
+  
+  button {
+    padding-bottom: 5px;
+    padding-top: 5px;
+    line-height: 160%;
+  }
+  
+  #home, #unit, #menu, #kc, #login, #logout, #profile {
+    margin-top: 20px;
+    font-size: 12px;
+    width: 85%;
+    margin-left: 7.5%;
+    white-space: normal;
+    text-align: center;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+    
+  .mat-drawer-inner-container::-webkit-scrollbar {
+    display: none;
+  }
+
