+++ src/app/modules/layout/navbar/navbar.component.html
@@ -1,23 +1,11 @@
 <div>
-    <span>
-      <img *ngIf="isDarkTheme" src="../../../../assets/images/logo.png" width="120px" alt="Clean CaDET" [routerLink]="['home']" (click)="resetNavBar()">
-      <img *ngIf="!isDarkTheme" src="../../../../assets/images/black.png" width="120px" alt="Clean CaDET" [routerLink]="['home'] " (click)="resetNavBar()">
-    </span>
-  <button color="primary" id="unit" mat-flat-button [disabled]="!learner" [matMenuTriggerFor]="unit">
-    {{selectedUnit ? selectedUnit.name : 'Izaberi lekciju'}}
-  </button>
-  <mat-menu #unit="matMenu">
-    <button [routerLink]="['unit', unit.id]" *ngFor="let unit of units" mat-menu-item>
-      {{unit.name}}
-    </button>
-  </mat-menu>
-  <button color="primary" id="kc" mat-flat-button [matMenuTriggerFor]="kc" [disabled]="!selectedUnit">
-    {{selectedKC ? selectedKC.name : 'Izaberi veštinu'}}
-  </button>
-  <mat-menu #kc="matMenu">
-    <cc-knowledge-map *ngIf="selectedUnit" [unitId]="selectedUnit.id" [knowledgeComponents]="selectedUnit.knowledgeComponents">
-    </cc-knowledge-map>
-  </mat-menu>
+  <span>
+    <img *ngIf="isDarkTheme" src="../../../../assets/images/logo.png" width="120px" alt="Clean CaDET" [routerLink]="['home']">
+    <img *ngIf="!isDarkTheme" src="../../../../assets/images/black.png" width="120px" alt="Clean CaDET" [routerLink]="['home']">
+  </span>
+
+  <cc-learner-controls [learner]="learner"></cc-learner-controls>
+  
   <button color="primary" *ngIf="!learner" id="login" mat-flat-button [routerLink]="['login']">Prijavi se</button>
   <p *ngIf="learner" color="primary" id="studentIndex">{{learner.index}}</p>
   <button color="primary" *ngIf="learner" mat-flat-button id="logout" (click)="onLogout()" [routerLink]="['home']">Odjavi se</button>
