+++ src/app/modules/layout/navbar/navbar.component.ts
@@ -16,99 +16,19 @@ import {InterfacingInstructor} from '../../instructor/interfacing-instructor.ser
 })
 export class NavbarComponent implements OnInit {
 
-  units: Unit[];
-  knowledgeComponents: KnowledgeComponent[];
   learner: Learner;
-  selectedUnit: Unit;
-  selectedKC: KnowledgeComponent;
   @Input() isDarkTheme: boolean;
 
-  constructor(private unitService: UnitService, private learnerService: LearnerService,
-              private router: Router, private route: ActivatedRoute,
-              private instructor: InterfacingInstructor) {
+  constructor(private learnerService: LearnerService) {
   }
 
   ngOnInit(): void {
     this.learnerService.learner$.subscribe(learner => {
-      if (learner == null) {
-        this.resetNavBar();
-        return;
-      }
       this.learner = learner;
-      this.updateUnits();
     });
-    this.instructor.observedAeEvaluations.subscribe(() => this.updateKnowledgeComponents(this.selectedUnit.id));
-    this.setupActiveUnitAndKCUpdate();
-  }
-
-  private updateUnits(): void {
-    this.unitService.getUnits().subscribe(units => {
-      this.units = units;
-    });
-  }
-
-  private setupActiveUnitAndKCUpdate(): void {
-    this.router.events.pipe(filter(e => e instanceof NavigationEnd),
-      map(e => this.getParams(this.route))
-    ).subscribe(params => {
-      if (this.unitIsChanged(params)) {
-        this.selectNewUnit(params);
-      } else if (params.kcId) {
-        this.selectedKC = this.findKC(this.selectedUnit.knowledgeComponents, +params.kcId);
-      }
-    });
-  }
-
-  private unitIsChanged(params: Params) {
-    return params.unitId && this.selectedUnit?.id != params.unitId;
-  }
-
-  private getParams(route: ActivatedRoute): Params {
-    let params = route.snapshot.params;
-    route.children?.forEach(c => {
-      params = {
-        ...this.getParams(c),
-        ...params
-      };
-    });
-    return params;
-  }
-
-  private selectNewUnit(params: Params): void {
-    this.unitService.getUnit(+params.unitId, this.learner.id).subscribe(fullUnit => {
-      this.knowledgeComponents = fullUnit.knowledgeComponents;
-      this.selectedUnit = fullUnit;
-      this.selectedKC = params.kcId ? this.findKC(this.selectedUnit.knowledgeComponents, +params.kcId) : null;
-    });
-  }
-
-  private updateKnowledgeComponents(unitId): void {
-    this.unitService.getUnit(unitId, this.learner.id).subscribe(fullUnit => {
-      this.selectedUnit = fullUnit;
-    });
-  }
-
-  private findKC(knowledgeComponents: KnowledgeComponent[], id: number): KnowledgeComponent {
-    for (const kc of knowledgeComponents) {
-      if (kc.id === id) {
-        return kc;
-      }
-
-      const child = this.findKC(kc.knowledgeComponents, id);
-      if (child) {
-        return child;
-      }
-    }
-    return null;
   }
 
   onLogout(): void {
     this.learnerService.logout();
-  }
-
-  resetNavBar(): void {
-    this.learner = null;
-    this.selectedUnit = null;
-    this.selectedKC = null;
   }
 }
