+++ src/app/modules/layout/navbar/navbar.module.ts
@@ -4,10 +4,12 @@ import { NavbarComponent } from './navbar.component';
 import { MaterialModule } from '../../../infrastructure/material.module';
 import { AppRoutingModule } from '../../../infrastructure/app-routing.module';
 import {DomainModule} from '../../domain/domain.module';
+import { LearnerControlsComponent } from './learner-controls/learner-controls.component';
 
 @NgModule({
   declarations: [
-    NavbarComponent
+    NavbarComponent,
+    LearnerControlsComponent
   ],
   imports: [
     BrowserModule,
