+++ src/app/modules/learner-analytics/kcm-progress/kcm-progress.component.html
@@ -1,6 +1,6 @@
-<div fxLayout="column" fxLayoutGap="20px" class="container-with-footer">
-    <div *ngFor="let p of progress">
-        <mat-card fxFlex="95" class="mat-elevation-z7">
+<div fxLayout="row wrap" class="container-with-footer">
+    <div *ngFor="let p of progress" fxFlex="48">
+        <mat-card class="mat-elevation-z7 margin">
             <mat-card-title>{{p.learner.name}} {{p.learner.surname}}</mat-card-title>
             <mat-card-subtitle>{{p.learner.index}}</mat-card-subtitle>
             <mat-card-content>
